requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
llama-parse>=0.4.0
openai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
pillow>=10.0.0
beautifulsoup4>=4.12.0
sentence-transformers>=2.2.0
transformers>=4.35.0
torch>=2.2.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
werkzeug>=2.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
nltk>=3.8.0
textstat>=0.7.0
python-pptx>=0.6.20
llama_cloud_services